separate getBooking in View
